% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{extdata}
\alias{extdata}
\title{External patternize data}
\format{
Raw JPG images, landmark and cartoon data.
\describe{
  \item{BC0077.JPG}{jpeg image}
  \item{BC0071.JPG}{jpeg image}
  \item{BC0050.JPG}{jpeg image}
  \item{BC0049.JPG}{jpeg image}
  \item{BC0004.JPG}{jpeg image}
  \item{BC0077_landmarks_LFW.Txt}{xy landmark coordinates}
  \item{BC0071_landmarks_LFW.Txt}{xy landmark coordinates}
  \item{BC0050_landmarks_LFW.Txt}{xy landmark coordinates}
  \item{BC0049_landmarks_LFW.Txt}{xy landmark coordinates}
  \item{BC0004_landmarks_LFW.Txt}{xy landmark coordinates}
  \item{BC0077_outline.txt}{xy outline coordinates}
  \item{BC0077_vein1.txt}{xy vein coordinates}
  \item{BC0077_vein2.txt}{xy vein coordinates}
  \item{BC0077_vein3.txt}{xy vein coordinates}
  \item{BC0077_vein4.txt}{xy vein coordinates}
  \item{BC0077_vein5.txt}{xy vein coordinates}
  \item{BC0077_vein6.txt}{xy vein coordinates}
  \item{BC0077_vein7.txt}{xy vein coordinates}
  \item{BC0077_vein8.txt}{xy vein coordinates}
  \item{BC0077_vein9.txt}{xy vein coordinates}
  \item{BC0077_vein10.txt}{xy vein coordinates}
  \item{BC0077_vein11.txt}{xy vein coordinates}
}
}
\description{
Raw image, landmark and cartoon data of Heliconius erato hydara wings.
}
