% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMImage.R
\name{GMMImage}
\alias{GMMImage}
\title{\code{\link[ClusterR]{GMM}} clustering of image imported as a RasterStack.}
\usage{
GMMImage(image, k = 5, maskToNA = NULL, kmeansOnAll = FALSE)
}
\arguments{
\item{image}{Image imported as a RasterStack for clustering.}

\item{k}{Integer for number of k clusters (default = 3).}

\item{maskToNA}{Replace the color value used for masking (i.e. 0 or 255) with NA.}

\item{kmeansOnAll}{Whether to perform the kmeans clusters on the combined set of pixels of all images
first (default = FALSE).}
}
\value{
List including the clustered \code{RasterSatck} returned as an array and object
   of class "\code{GMM}".
}
\description{
\code{\link[ClusterR]{GMM}} clustering of image imported as a RasterStack.
}
