% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imageList}
\alias{imageList}
\title{imageList}
\format{
A list of 5 \code{RasterStack} objects of Heliconius erato hydara dorsal forewings.
}
\usage{
imageList
}
\description{
List of RasterStacks as returned by \code{makeList}.
}
\examples{
\dontrun{
data(imageList)
summary(imageList)
}
}
\keyword{datasets}
