% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kImage.R
\name{kImage}
\alias{kImage}
\title{kmeans clustering of image.}
\usage{
kImage(image, k = 5, startCenter = NULL)
}
\arguments{
\item{image}{Image for kmeans clustering.}

\item{k}{Integer for number of kmeans clusters (default = 3).}

\item{startCenter}{A matrix of cluster centres to start kmeans clustering from (default = NULL).}
}
\value{
List with kmeans clustered \code{image} and object of class "kmeans".
}
\description{
kmeans clustering of image.
}
\examples{
image <- raster::stack(system.file("extdata", "BC0077.jpg", package = "patternize"))
out <- kImage(image, 6)

}

