% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanArray.R
\name{lanArray}
\alias{lanArray}
\title{Build landmark array for Morpho.}
\usage{
lanArray(sampleList, adjustLandmarks = F)
}
\arguments{
\item{sampleList}{List of landmark matrices.}

\item{adjustLandmarks}{Adjust the coordinates of the lanmarks (default = FALSE).}
}
\value{
X x Y x n array, where X and Y define the coordinates of the landmark points and n is the sample size.
}
\description{
Build landmark array for Morpho.
}
\examples{
data(landmarks)
landmarkArray <- lanArray(landmarks)

}

