% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landmarkArray}
\alias{landmarkArray}
\title{landmarkArray}
\format{
An array of landmarks for 5 Heliconius erato hydara dorsal forewings.
}
\usage{
landmarkArray
}
\description{
Array of landmarks as returned by \code{lanArray} and used by
   \code{link[Morpho]{procsym}}.
}
\examples{
\dontrun{
data(landmarkArray)
summary(landmarkArray)
}
}
\keyword{datasets}
