% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landmarkList}
\alias{landmarkList}
\title{landmarkList}
\format{
A list of landmarks for 5 Heliconius erato hydara dorsal forewings.
}
\usage{
landmarkList
}
\description{
List of landmarks as returned by \code{makeList}.
}
\examples{
\dontrun{
data(landmarkList)
summary(landmarkList)
}
}
\keyword{datasets}
