% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanList.R
\name{lanList}
\alias{lanList}
\title{Build landmark list using filepath and fileextension.}
\usage{
lanList(IDlist, prepath = NULL, extension = NULL)
}
\arguments{
\item{IDlist}{List of sample IDs.}

\item{prepath}{Prepath (default = NULL).}

\item{extension}{Extension (default = NULL).}
}
\value{
Landmark list.
}
\description{
Build landmark list using filepath and fileextension.
}
\examples{
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
prepath <- system.file("extdata",  package = 'patternize')
extension <- '_landmarks_LFW.txt'
landmarkList <- lanList(IDlist, prepath, extension)

}

