% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patArea.R
\name{patArea}
\alias{patArea}
\title{Calculate relative area of color pattern}
\usage{
patArea(rList, IDlist, refShape, type, outline = NULL, landList = NULL,
  adjustCoords = FALSE, cartoonID = NULL, crop = c(0, 0, 0, 0),
  flipRaster = NULL, flipOutline = NULL, imageList = NULL)
}
\arguments{
\item{rList}{List of raster objects.}

\item{IDlist}{List of sample IDs.}

\item{refShape}{This can be 'target' in case the reference shape is a single sample (for registration analysis) or 'mean' if the images were transformed to a mean shape (only for meanshape when using landmark transformation)}

\item{outline}{xy coordinates that define outline.}

\item{landList}{Landmark landmarkList.}

\item{adjustCoords}{Adjust landmark coordinates.}

\item{cartoonID}{ID of the sample for which the cartoon was drawn.}

\item{crop}{Vector c(xmin, xmax, ymin, ymax) that specifies the pixel coordinates to crop the original image used in landmark or registration analysis.}

\item{flipRaster}{Whether to flip raster along xy axis (in case there is an inconsistency between raster and outline coordinates).}

\item{flipOutline}{Whether to flip plot along x, y or xy axis.}

\item{imageList}{List of image should be given if one wants to flip the outline or adjust landmark coordinates.}
}
\description{
Calculate relative area of color pattern
}
\examples{
data(rasterList_lanRGB)
data(rasterList_regRGB)
data(rasterList_lanK)
data(rasterList_regK)

data(imageList)

IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
outline_BC0077 <- read.table(paste(system.file("extdata",  package = 'patternize'), '/BC0077_outline.txt', sep=''), h= F)
prepath <- system.file("extdata", package = 'patternize')
extension <- '_landmarks_LFW.txt'
landmarkList <- makeList(IDlist, 'landmark', prepath, extension)
summedRaster_lanRGB <- sumRaster(rasterList_lanRGB, IDlist, type = 'RGB')

patArea(rasterList_lanRGB, IDlist, refShape = 'mean', type = 'RGB', outline = outline_BC0077, landList = landmarkList, adjustCoords = TRUE, imageList = imageList, cartoonID = 'BC0077')
patArea(rasterList_regRGB, IDlist, refShape = 'target', type = 'RGB', outline = outline_BC0077,  crop = c(1000,4000,400,2500), adjustCoords = TRUE, imageList = imageList, cartoonID = 'BC0077', flipRaster = 'xy')
patArea(rasterList_lanK, IDlist, refShape = 'mean', type = 'k', outline = outline_BC0077, landList = landmarkList, adjustCoords = TRUE, imageList = imageList, cartoonID = 'BC0077')
patArea(rasterList_regK, IDlist, refShape = 'target', type = 'k', outline = outline_BC0077,  crop = c(1000,4000,400,2500), adjustCoords = TRUE, imageList = imageList, cartoonID = 'BC0077', flipRaster = 'xy')

}

