% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patLanK.R
\name{patLanK}
\alias{patLanK}
\title{Color pattern quantification using landmarks and k-means clustering for color extraction.}
\usage{
patLanK(imageList, landmarkList, k = 3, resampleFactor = 1, crop = FALSE,
  cropOffset = NULL, res = 300, transformRef = "meanshape",
  transformType = "tps", removebg = FALSE, adjustCoords = FALSE,
  plot = FALSE, focal = FALSE, sigma = 3)
}
\arguments{
\item{imageList}{List of RasterStack objects.}

\item{k}{Integere for defining number of k-means clusters (default = 3).}

\item{resampleFactor}{Integer for downsampling used by \code{\link{redRes}}.}

\item{crop}{Whether to use the landmarks range to crop the image. This can significantly speed up the analysis (default = FALSE).}

\item{cropOffset}{Vector c(xmin, xmax, ymin, ymax) that specifies the number of pixels you want the cropping to be offset from the landmarks (in case the landmarks do not surround the entire color pattern).}

\item{res}{Resolution for color pattern raster (default = 300). This should be reduced if the number of pixels in the image is lower than th raster.}

\item{transformRef}{ID of reference sample for shape to which color patterns will be transformed to. Can be 'meanshape' for transforming to mean shape of Procrustes analysis.}

\item{transformType}{(default ='tps')}

\item{removebg}{Whether to remove white background rasterList (default = FALSE) for k-means analysis.}

\item{adjustCoords}{Adjust coordinates.}

\item{plot}{Whether to plot transformed color patterns while processing (default = FALSE).}

\item{focal}{Whether to perform Gaussian blurring (default = FALSE).}

\item{sigma}{Size of sigma for Gaussian blurring (default = 3).}

\item{lanArray}{Landmark array.}
}
\value{
List of summed raster for each k-means cluster objects.
}
\description{
Color pattern quantification using landmarks and k-means clustering for color extraction.
}
\examples{
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
prepath <- system.file("extdata",  package = 'patternize')
extension <- '_landmarks_LFW.txt'
landmarkList <- makeList(IDlist, 'landmark', prepath, extension)

extension <- '.JPG'
imageList <- makeList(IDlist, 'image', prepath, extension)

rasterList_lanK <- patLanK(imageList, landmarkList, k = 5, resampleFactor = 3, crop = TRUE, res = 150, removebg = TRUE, adjustCoords = TRUE, plot = TRUE)

}

