% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patPCA.R
\name{patPCA}
\alias{patPCA}
\title{Performs PCA}
\usage{
patPCA(rList, popList, colList)
}
\arguments{
\item{rList}{List of raster objects.}

\item{popList}{List of vectors including sampleIDs for eacht population.}

\item{colList}{List of colors for each population.}
}
\description{
Performs PCA
}
\examples{
data(rasterList_lanRGB)

pop1 <- c('BC0077','BC0071')
pop2 <- c('BC0050','BC0049','BC0004')
popList <- list(pop1, pop2)
colList <- c("red", "blue")

pcaOut <- patPCA(rasterList_lanRGB, popList, colList)
comp <- prcomp(pcaOut[[1]])
plot(comp$rotation, col=pcaOut[[2]]$col, pch=19)

}
