% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRegK.R
\name{patRegK}
\alias{patRegK}
\title{Image registration with k-means clustering.}
\usage{
patRegK(sampleList, target, k = 3, resampleFactor = 1,
  useBlockPercentage = 75, removebg = FALSE, plot = FALSE)
}
\arguments{
\item{sampleList}{List of RasterStack objects.}

\item{target}{Image used as target for registration.}

\item{k}{Integere for defining number of k-means clusters (default = 3).}

\item{resampleFactor}{Integer for downsampling used by \code{\link{redRes}}.}

\item{useBlockPercentage}{Block percentage as used in NiftyReg.}

\item{removebg}{Whether to remove white background rasterList (default = FALSE).}

\item{plot}{Whether to plot k-means clustered image while processing (default = FALSE).}
}
\value{
List of summed raster for each k-means cluster objects.
}
\description{
Image registration with k-means clustering.
}
\examples{
data(pictures)
target <- pictures[[1]]
rasterList <- patRegK(pictures, target, k = 6, removebg = TRUE, plot = TRUE)

}

