% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRegRGB.R
\name{patRegRGB}
\alias{patRegRGB}
\title{Image registration with RGB color extraction.}
\usage{
patRegRGB(sampleList, target, RGB, resampleFactor = 1,
  useBlockPercentage = 75, colOffset = 0, removebg = FALSE,
  plot = FALSE)
}
\arguments{
\item{sampleList}{List of RasterStack objects.}

\item{target}{Image used as target for registration.}

\item{RGB}{RGB values for color pattern extraction specified as vector.}

\item{resampleFactor}{Integer for downsampling used by \code{\link{redRes}}.}

\item{useBlockPercentage}{Block percentage as used in NiftyReg.}

\item{colOffset}{Color offset for color pattern extraction (default = 0).}

\item{removebg}{Whether to remove white background for image registration (default = FALSE).}

\item{plot}{Whether to plot transformed color patterns while processing (default = FALSE).}
}
\value{
List of raster objects.
}
\description{
Image registration with RGB color extraction.
}
\examples{
data(pictures)
target <- pictures[[1]]
RGB <- c(114,17,0)
rasterList <- patReg(pictures, target, RGB, colOffset= 0.15, removebg = TRUE, plot = TRUE)

}

