% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternize.R
\docType{package}
\name{patternize}
\alias{patternize}
\alias{patternize-package}
\title{patternize: An R package for quantifying color pattern variation.}
\description{
The patternize package provides two main functions:
patCol and patLan.
}
\section{patternize functions}{

The patternize functions ...
}

