% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRasterStackAsImage.R
\name{plotRasterstackAsImage}
\alias{plotRasterstackAsImage}
\title{Plot rasterStack as image.}
\usage{
plotRasterstackAsImage(rasterStack)
}
\arguments{
\item{rasterStack}{A single rasterStack.}
}
\description{
Plot rasterStack as image.
}
