% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rasterList_lanK}
\alias{rasterList_lanK}
\title{rasterList_lanK}
\format{A list of RasterLayers including the red color pattern extracted from 5 Heliconius
   erato hydara dorsal forewings using \code{patLanK}.}
\usage{
rasterList_lanK
}
\description{
List of RasterLayers as returned by \code{patLanK}.
}
\examples{
\dontrun{
data(rasterList_lanK)
summary(rasterList_lanL)
}
}
\keyword{datasets}
