% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rasterList_lanRGB}
\alias{rasterList_lanRGB}
\title{rasterList_lanRGB}
\format{A list of RasterLayers including the red color pattern extracted from 5 Heliconius
    erato hydara dorsal forewings using \code{patLanRGB}.}
\usage{
rasterList_lanRGB
}
\description{
List of RasterLayers as returned by \code{patLanRGB}.
}
\examples{
\dontrun{
data(rasterList_lanRGB)
summary(rasterList_lanRGB)
}
}
\keyword{datasets}
