% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rasterList_regK}
\alias{rasterList_regK}
\title{rasterList_regK}
\format{A list of RasterLayers including the red color pattern extracted from 5 Heliconius
   erato hydara dorsal forewings using \code{patRegK}.}
\usage{
rasterList_regK
}
\description{
List of RasterLayers as returned by \code{patRegK}.
}
\examples{
\dontrun{
data(rasterList_regK)
summary(rasterList_regK)
}
}
\keyword{datasets}
