% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rasterList_regRGB}
\alias{rasterList_regRGB}
\title{rasterList_regRGB}
\format{A list of RasterLayers including the red color pattern extracted from 5 Heliconius
   erato hydara dorsal forewings using \code{patRegRGB}.}
\usage{
rasterList_regRGB
}
\description{
List of RasterLayers as returned by \code{patRegRGB}.
}
\examples{
\dontrun{
data(rasterList_regRGB)
summary(rasterList_regRGB)
}
}
\keyword{datasets}
