% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redRes.R
\name{redRes}
\alias{redRes}
\title{Reduce the resolution of an iamge.}
\usage{
redRes(x, r)
}
\arguments{
\item{x}{image.}

\item{r}{resamplefactor.}
}
\value{
Reduce \code{x}.
}
\description{
Reduce the resolution of an iamge.
}
\examples{
redRes(x, 10)
}

