% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMask.R
\name{setMask}
\alias{setMask}
\title{Interactive function to to draw an outline for masking.}
\usage{
setMask(summedRaster, IDlist, filename, ...)
}
\arguments{
\item{summedRaster}{Summed raster of extracted patterns.}

\item{IDlist}{List of sample IDs.}

\item{filename}{Name of file to which mask will be written.}

\item{...}{additional arguments for plotHeat function.}
}
\value{
file
}
\description{
Interactive function to to draw an outline for masking.
}
