% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumRaster.R
\name{sumRaster}
\alias{sumRaster}
\title{sum rasterlist for set of samples}
\usage{
sumRaster(rList, IDlist, type)
}
\arguments{
\item{rList}{List of raster objects.}

\item{IDlist}{List of sample IDs.}

\item{type}{Type of rasterlist; simple (RGB) or nested (k) (result from RGB or k-means analysis, respectively)}
}
\description{
sum rasterlist for set of samples
}
\examples{
data(rasterList_lanRGB)
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
summedRaster <- sumRaster(rasterList_lanRGB, IDlist, type = 'RGB')

data(rasterList_lanK)
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
summedRasterList <- sumRaster(rasterList_lanK, IDlist, type = 'k')

}

