% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorChecker.R
\name{colorChecker}
\alias{colorChecker}
\title{Calibrate images using ColorChecker.}
\usage{
colorChecker(IDlist, prepath = NULL, extension = NULL,
  patchSize = 0.6)
}
\arguments{
\item{IDlist}{List of sample IDs.}

\item{prepath}{Prepath (default = NULL).}

\item{extension}{Extension (default = NULL).}

\item{patchSize}{Proportion of ColorChecker patch that will be used for observed RGB values
(default = NULL).}
}
\value{
Calibrated image(s) ('filename_calibrated.jpg')
}
\description{
Calibrate images using ColorChecker.
}
