% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanArray.R
\name{lanArray}
\alias{lanArray}
\title{Build landmark array for Morpho.}
\usage{
lanArray(sampleList, adjustCoords = FALSE, image = NULL)
}
\arguments{
\item{sampleList}{List of landmark matrices.}

\item{adjustCoords}{Adjust the coordinates of the landmarks if different (default = FALSE).}

\item{image}{Image List.}
}
\value{
X x Y x n array, where X and Y define the coordinates of the landmark points and n is the sample size.
}
\description{
Build landmark array for Morpho.
}
\examples{
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
prepath <- system.file("extdata",  package = 'patternize')
extension <- '_landmarks_LFW.txt'
landmarkList <- makeList(IDlist, 'landmark', prepath, extension)

landmarkArray <- lanArray(landmarkList)

}
