% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRegK.R
\name{patRegK}
\alias{patRegK}
\title{Image registration with k-means clustering for color extraction.}
\usage{
patRegK(sampleList, target, k = 3, resampleFactor = 1,
  useBlockPercentage = 75, crop = NULL, removebg = FALSE, plot = FALSE,
  focal = FALSE, sigma = 3)
}
\arguments{
\item{sampleList}{List of RasterStack objects.}

\item{target}{Image used as target for registration.}

\item{k}{Integere for defining number of k-means clusters (default = 3).}

\item{resampleFactor}{Integer for downsampling used by \code{\link{redRes}}.}

\item{useBlockPercentage}{Block percentage as used in NiftyReg.}

\item{crop}{Vector c(xmin, xmax, ymin, ymax) that specifies the pixel coordinates to crop the original image.}

\item{removebg}{Whether to remove white background rasterList (default = FALSE) for registration and k-means analysis.}

\item{plot}{Whether to plot k-means clustered image while processing (default = FALSE).}

\item{focal}{Whether to perform Gaussian blurring (default = FALSE).}

\item{sigma}{Size of sigma for Gaussian blurring (default = 3).}
}
\value{
List of summed raster for each k-means cluster objects.
}
\description{
Image registration with k-means clustering for color extraction.
}
\examples{
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
prepath <- system.file("extdata",  package = 'patternize')
extension <- '.jpg'
imageList <- makeList(IDlist, 'image', prepath, extension)
target <- imageList[[1]]
rasterList_regK <- patRegK(imageList, target, k = 5, resampleFactor = 10, crop = c(1000,4000,400,2500), removebg = TRUE, plot = TRUE)

}

