% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patRegRGB.R
\name{patRegRGB}
\alias{patRegRGB}
\title{Image registration with RGB color extraction.}
\usage{
patRegRGB(sampleList, target, RGB, resampleFactor = 1,
  useBlockPercentage = 75, colOffset = 0, crop = NULL, removebg = FALSE,
  plot = FALSE, focal = FALSE, sigma = 3, iterations = 0)
}
\arguments{
\item{sampleList}{List of RasterStack objects.}

\item{target}{Image used as target for registration.}

\item{RGB}{RGB values for color pattern extraction specified as vector.}

\item{resampleFactor}{Integer for downsampling used by \code{\link{redRes}}.}

\item{useBlockPercentage}{Block percentage as used in NiftyReg.}

\item{colOffset}{Color offset for color pattern extraction (default = 0).}

\item{crop}{Vector c(xmin, xmax, ymin, ymax) that specifies the pixel coordinates to crop the original image.}

\item{removebg}{Whether to remove white background for image registration (default = FALSE).}

\item{plot}{Whether to plot transformed color patterns while processing (default = FALSE).}

\item{focal}{Whether to perform Gaussian blurring (default = FALSE).}

\item{sigma}{Size of sigma for Gaussian blurring (default = 3).}

\item{iterations}{Number of iterations for recalculating average color.}
}
\value{
List of raster objects.
}
\description{
Image registration with RGB color extraction.
}
\examples{
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
prepath <- system.file("extdata",  package = 'patternize')
extension <- '.jpg'
imageList <- makeList(IDlist, 'image', prepath, extension)
target <- imageList[[1]]
RGB <- c(114,17,0)
rasterList_regRGB <- patRegRGB(imageList, target, RGB, resampleFactor = 10, colOffset= 0.15, crop = c(1000,4000,400,2500), removebg = TRUE, plot = TRUE)

}

