% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeat.R
\name{plotHeat}
\alias{plotHeat}
\title{Plot heatmap from summed rasterList}
\usage{
plotHeat(summedRaster, IDlist, colpalette = NULL)
}
\arguments{
\item{summedRaster}{Summed raster or summedRasterList.}

\item{IDlist}{List of sample IDs.}

\item{colpalette}{Vector of colors for color palette (default = c("white","lightblue","blue","green", "yellow","red"))}
}
\description{
Plot heatmap from summed rasterList
}
\examples{
data(rasterList_lanRGB)
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
summedRaster <- sumRaster(rasterList_lanRGB, IDlist, type = 'RGB')

plotHeat(summedRaster, IDlist)

data(rasterList_lanK)
IDlist <- c('BC0077','BC0071','BC0050','BC0049','BC0004')
summedRasterList <- sumRaster(rasterList_lanK, IDlist, type = 'k')

plotHeat(summedRasterList, IDlist)

plotHeat(summedRasterList[[2]], IDlist)

}

