% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRasterStackAsImage.R
\name{plotRasterstackAsImage}
\alias{plotRasterstackAsImage}
\title{Plot rasterStack as image.}
\usage{
plotRasterstackAsImage(rasterStack, flipY = FALSE)
}
\arguments{
\item{rasterStack}{A single rasterStack.}

\item{flipY}{Whether to flip the raster along the Y-axis.}
}
\description{
Plot rasterStack as image.
}
