% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redRes.R
\name{redRes}
\alias{redRes}
\title{Reduce the resolution of an iamge by downsampling.}
\usage{
redRes(image, resampleFactor)
}
\arguments{
\item{image}{Image for downsampling.}

\item{resampleFactor}{Integer for downsampling.}
}
\value{
Downsampled \code{image}.
}
\description{
Reduce the resolution of an iamge by downsampling.
}
\examples{
image <- raster::stack(system.file("extdata", "BC0077.jpg", package = "patternize"))
image.reduced <- redRes(image, 5)

}
